feat: fullscreen rendering
